chore: add dev-utils
