+++ src/dev-utils/debug.ts
@@ -1,0 +1,5 @@
+export { isInDebug };
+
+const isInDebug = (): boolean => {
+  return !process.env.NODE_ENV || process.env.NODE_ENV === "development";
+};
